+++ monai/transforms/post/dictionary.py
@@ -39,7 +39,7 @@ from monai.transforms.post.array import (
 from monai.transforms.transform import MapTransform
 from monai.transforms.utility.array import ToTensor
 from monai.transforms.utils import allow_missing_keys_mode, convert_inverse_interp_mode
-from monai.utils import ensure_tuple, ensure_tuple_rep
+from monai.utils import deprecated_arg, ensure_tuple, ensure_tuple_rep
 from monai.utils.enums import InverseKeys
 
 __all__ = [
@@ -126,16 +126,18 @@ class AsDiscreted(MapTransform):
     Dictionary-based wrapper of :py:class:`monai.transforms.AsDiscrete`.
     """
 
+    @deprecated_arg("n_classes", since="0.6")
     def __init__(
         self,
         keys: KeysCollection,
         argmax: Union[Sequence[bool], bool] = False,
         to_onehot: Union[Sequence[bool], bool] = False,
-        n_classes: Optional[Union[Sequence[int], int]] = None,
+        num_classes: Optional[Union[Sequence[int], int]] = None,
         threshold_values: Union[Sequence[bool], bool] = False,
         logit_thresh: Union[Sequence[float], float] = 0.5,
         rounding: Union[Sequence[Optional[str]], Optional[str]] = None,
         allow_missing_keys: bool = False,
+        n_classes: Optional[int] = None,
     ) -> None:
         """
         Args:
@@ -145,7 +147,7 @@ class AsDiscreted(MapTransform):
                 it also can be a sequence of bool, each element corresponds to a key in ``keys``.
             to_onehot: whether to convert input data into the one-hot format. Defaults to False.
                 it also can be a sequence of bool, each element corresponds to a key in ``keys``.
-            n_classes: the number of classes to convert to One-Hot format. it also can be a
+            num_classes: the number of classes to convert to One-Hot format. it also can be a
                 sequence of int, each element corresponds to a key in ``keys``.
             threshold_values: whether threshold the float value to int number 0 or 1, default is False.
                 it also can be a sequence of bool, each element corresponds to a key in ``keys``.
@@ -157,10 +159,13 @@ class AsDiscreted(MapTransform):
             allow_missing_keys: don't raise exception if key is missing.
 
         """
+        # in case the new num_classes is default but you still call deprecated n_classes
+        if n_classes is not None and num_classes is None:
+            num_classes = n_classes
         super().__init__(keys, allow_missing_keys)
         self.argmax = ensure_tuple_rep(argmax, len(self.keys))
         self.to_onehot = ensure_tuple_rep(to_onehot, len(self.keys))
-        self.n_classes = ensure_tuple_rep(n_classes, len(self.keys))
+        self.num_classes = ensure_tuple_rep(num_classes, len(self.keys))
         self.threshold_values = ensure_tuple_rep(threshold_values, len(self.keys))
         self.logit_thresh = ensure_tuple_rep(logit_thresh, len(self.keys))
         self.rounding = ensure_tuple_rep(rounding, len(self.keys))
@@ -168,14 +173,14 @@ class AsDiscreted(MapTransform):
 
     def __call__(self, data: Mapping[Hashable, torch.Tensor]) -> Dict[Hashable, torch.Tensor]:
         d = dict(data)
-        for key, argmax, to_onehot, n_classes, threshold_values, logit_thresh, rounding in self.key_iterator(
-            d, self.argmax, self.to_onehot, self.n_classes, self.threshold_values, self.logit_thresh, self.rounding
+        for key, argmax, to_onehot, num_classes, threshold_values, logit_thresh, rounding in self.key_iterator(
+            d, self.argmax, self.to_onehot, self.num_classes, self.threshold_values, self.logit_thresh, self.rounding
         ):
             d[key] = self.converter(
                 d[key],
                 argmax,
                 to_onehot,
-                n_classes,
+                num_classes,
                 threshold_values,
                 logit_thresh,
                 rounding,