+++ tests/test_compute_roc_auc.py
@@ -87,7 +87,7 @@ class TestComputeROCAUC(unittest.TestCase):
     @parameterized.expand([TEST_CASE_1, TEST_CASE_2, TEST_CASE_3, TEST_CASE_4, TEST_CASE_5, TEST_CASE_6, TEST_CASE_7])
     def test_value(self, y_pred, y, softmax, to_onehot, average, expected_value):
         y_pred_trans = Compose([ToTensor(), Activations(softmax=softmax)])
-        y_trans = Compose([ToTensor(), AsDiscrete(to_onehot=to_onehot, n_classes=2)])
+        y_trans = Compose([ToTensor(), AsDiscrete(to_onehot=to_onehot, num_classes=2)])
         y_pred = torch.stack([y_pred_trans(i) for i in decollate_batch(y_pred)], dim=0)
         y = torch.stack([y_trans(i) for i in decollate_batch(y)], dim=0)
         result = compute_roc_auc(y_pred=y_pred, y=y, average=average)
@@ -96,7 +96,7 @@ class TestComputeROCAUC(unittest.TestCase):
     @parameterized.expand([TEST_CASE_1, TEST_CASE_2, TEST_CASE_3, TEST_CASE_4, TEST_CASE_5, TEST_CASE_6, TEST_CASE_7])
     def test_class_value(self, y_pred, y, softmax, to_onehot, average, expected_value):
         y_pred_trans = Compose([ToTensor(), Activations(softmax=softmax)])
-        y_trans = Compose([ToTensor(), AsDiscrete(to_onehot=to_onehot, n_classes=2)])
+        y_trans = Compose([ToTensor(), AsDiscrete(to_onehot=to_onehot, num_classes=2)])
         y_pred = [y_pred_trans(i) for i in decollate_batch(y_pred)]
         y = [y_trans(i) for i in decollate_batch(y)]
         metric = ROCAUCMetric(average=average)