+++ tests/test_torchvision_fc_model.py
@@ -24,19 +24,19 @@ _, has_tv = optional_import("torchvision")
 device = "cuda" if torch.cuda.is_available() else "cpu"
 
 TEST_CASE_0 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": True, "pretrained": False},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": True, "pretrained": False},
     (2, 3, 224, 224),
     (2, 1, 1, 1),
 ]
 
 TEST_CASE_1 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": True, "pretrained": False},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": True, "pretrained": False},
     (2, 3, 256, 256),
     (2, 1, 2, 2),
 ]
 
 TEST_CASE_2 = [
-    {"model_name": "resnet101", "n_classes": 5, "use_conv": True, "pretrained": False},
+    {"model_name": "resnet101", "num_classes": 5, "use_conv": True, "pretrained": False},
     (2, 3, 256, 256),
     (2, 5, 2, 2),
 ]
@@ -44,7 +44,7 @@ TEST_CASE_2 = [
 TEST_CASE_3 = [
     {
         "model_name": "resnet101",
-        "n_classes": 5,
+        "num_classes": 5,
         "use_conv": True,
         "pool": ("avg", {"kernel_size": 6, "stride": 1}),
         "pretrained": False,
@@ -54,60 +54,60 @@ TEST_CASE_3 = [
 ]
 
 TEST_CASE_4 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": False, "pool": None, "pretrained": False},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": False, "pool": None, "pretrained": False},
     (2, 3, 224, 224),
     (2, 1),
 ]
 
 TEST_CASE_5 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": False, "pool": None, "pretrained": False},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": False, "pool": None, "pretrained": False},
     (2, 3, 256, 256),
     (2, 1),
 ]
 
 TEST_CASE_6 = [
-    {"model_name": "resnet101", "n_classes": 5, "use_conv": False, "pool": None, "pretrained": False},
+    {"model_name": "resnet101", "num_classes": 5, "use_conv": False, "pool": None, "pretrained": False},
     (2, 3, 256, 256),
     (2, 5),
 ]
 
 TEST_CASE_PRETRAINED_0 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": True, "pretrained": True},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": True, "pretrained": True},
     (2, 3, 224, 224),
     (2, 1, 1, 1),
     -0.010419349186122417,
 ]
 
 TEST_CASE_PRETRAINED_1 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": True, "pretrained": True},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": True, "pretrained": True},
     (2, 3, 256, 256),
     (2, 1, 2, 2),
     -0.010419349186122417,
 ]
 
 TEST_CASE_PRETRAINED_2 = [
-    {"model_name": "resnet18", "n_classes": 5, "use_conv": True, "pretrained": True},
+    {"model_name": "resnet18", "num_classes": 5, "use_conv": True, "pretrained": True},
     (2, 3, 256, 256),
     (2, 5, 2, 2),
     -0.010419349186122417,
 ]
 
 TEST_CASE_PRETRAINED_3 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": False, "pool": None, "pretrained": True},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": False, "pool": None, "pretrained": True},
     (2, 3, 224, 224),
     (2, 1),
     -0.010419349186122417,
 ]
 
 TEST_CASE_PRETRAINED_4 = [
-    {"model_name": "resnet18", "n_classes": 1, "use_conv": False, "pool": None, "pretrained": True},
+    {"model_name": "resnet18", "num_classes": 1, "use_conv": False, "pool": None, "pretrained": True},
     (2, 3, 256, 256),
     (2, 1),
     -0.010419349186122417,
 ]
 
 TEST_CASE_PRETRAINED_5 = [
-    {"model_name": "resnet18", "n_classes": 5, "use_conv": False, "pool": None, "pretrained": True},
+    {"model_name": "resnet18", "num_classes": 5, "use_conv": False, "pool": None, "pretrained": True},
     (2, 3, 256, 256),
     (2, 5),
     -0.010419349186122417,