+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* renamed model's `n_classes` to `num_classes`
+
 ## [0.6.0] - 2021-07-08
 ### Added
 * 10 new transforms, a masked loss wrapper, and a `NetAdapter` for transfer learning